requirements.txt
asgiref==3.7.2
Django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
gunicorn==21.2.0
Markdown==3.4.3
packaging==23.1
Pillow==10.0.0
pytz==2023.3
sqlparse==0.4.4
